+++ android/src/main/java/io/snapcall/reactnative/RNSnapcallReactModule.java
@@ -4,9 +4,11 @@ package io.snapcall.reactnative;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
 import android.util.Log;
+import androidx.annotation.Nullable;
+import androidx.annotation.NonNull;
+
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;